refactor: auto load routes across bffs
