+++ apps/bff-player/app/routes/health.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+@router.get("/player/healthz")
+async def health():
+    return {"status": "ok"}
