+++ gateway/realtime/app/routes/health.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+from .ws import redis
+
+router = APIRouter()
+
+
+@router.get("/healthz")
+async def health():
+    pong = await redis.ping()
+    return {"status": "ok", "redis": pong}
